+++ medily-backend/src/main/java/com/medily/backend/dto/prescriptionrequest/PrescriptionRequestCreateDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class PrescriptionRequestCreateDTO {
     @NotNull
-    private Long prescriptionId;
+    private Integer prescriptionId;
     @NotNull
-    private Long pharmacyId;
+    private Integer pharmacyId;
 }
