+++ medily-backend/src/main/java/com/medily/backend/repository/AppointmentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
+public interface AppointmentRepository extends JpaRepository<Appointment, Integer> {
     List<Appointment> findByPatientPatientId(Integer patientId);
     List<Appointment> findByDoctorDoctorId(Integer doctorId);
     List<Appointment> findByPatientPatientIdAndStatus(Integer patientId, Appointment.Status status);
